requirements.txt
flask==3.0.3
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
flask-socketio==5.3.6
flask-cors==4.0.1
flask-mail==0.10.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
